requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
requests==2.31.0
Flask-Session==0.5.0
Pillow==10.0.0
